feat: rework parser to produce AST
Instead of parsing the PDF into a document like data structure, the
parser now produces an AST. This approach results in cleaner code and a
more robust parser.

Currently the performance isn't great for larger PDFs, mainly due to how
stream data is handled (probably better to return the bytes directly
instead of a stringified version, or a reader?). The tokeniser is the
source of these issues. It handles stream data slowly and wastes memory
storing all read/unread tokens which causes unnecessary reallocation of
the arrays that back them (use a fixed size array of 2 or 3? Need to
confirm if this will be enough in all situations). These issues will be
addressed in the future, once the parser is more stable.