fix: improve parsing memory usage
Memory usage during parsing could grow unbounded. This change introduces
an LRU stack to limit the number of read/unread tokens that are held in
memory while parsing. The stack based provides a nice interface while
parsing (as it fits the read/unread process), and also avoids the need
for constantly growing the array after a push operation. This could
could've been solved by just limiting the size of the array with
slicing, but instead the technical stuff is hidden inside the stack
implementation.

Currently the stack size is limited to 3 for both read and unread
tokens. This should be okay for all situations as the parser only looks
back 2 tokens at most.

For small PDFs, the improvement is minimal, but for larger PDFs the
memory usage is significantly reduced.