fix: parsing hanging on stream bodies
Parser could hang indefinitely when parsing stream bodies (particularly
those that contain delimiter characters).

The tokeniser would return a KEYWORD type token with an empty string
value when it encountered a delimiter string in a stream body. This is
because of how the tokeniser defaults to a KEYWORD type token if it
doesn't match any other token type and the subsequent attempted reading
of "regular characters" (as per the PDF spec) which stops at whitespace
or delimiter characters. After the delimiter is hit, the tokeniser would
unread the last character and return an empty string.

The parser would continually ask for the next token from the tokeniser
but would receive a malformed KEYWORD token which ignore and the loop
would continue indefinitely.

The current fix is hacky, but works. A more elegant solution instead of
bulldozing through characters until we reach the endstream keyword would
be better.