feat: implement (somewhat working) serialiser
Implement a serialiser to turn an AST into a PDF. The serialised PDF
won't open in a viewer (it'll say it's damaged) because the xref table
is incorrect and there is no trailer present. It also has a hardcoded
PDF version. The objects forming the body of the PDF look to be correct.

Things wrong with the xref table:
- Entries are all on the same line (incorrect format)
- The byte offset for each entry is not padded to the required length
- Even if the byte offsets were correctly padded, they'd be wrong
- The generation number is not padded to the required length
- There is no item indicating if the entry is free or in use

Anything relating to byte offsets (like each entry in the xref table and
the startxref value) will be incorrect because the AST produced by the
parser doesn't perfectly preserve the original PDF, so there's a good
chance the serialised PDF will be different to the original.

If you were to build the AST manually and then serialise it, it'd be
impossible (technically possible but it'd take a very long time and
would rely on knowing exactly how the serialiser will output objects) to
create an accurate xref table because you wouldn't be able to know the
byte offsets. To fix this, the serialiser will need to create the xref
table (and startxref value) during serialisation.

Creating the xref table during serialisation also allows us to skip
parsing the xref table, which can be complicated if there have been
incremental updates to the PDF. There can be multiple trailers in an
incrementally updated PDF, (which might be difficult?) but can tackle
that later.