feat: generate xref table during serialisation
a28d83c added the serialiser but it wouldn't produce a valid PDF (as the
xref table had the wrong offsets and it was incorrectly formatted). This
commit produces a correct xref table and xref table offset.

The parser/serialiser combo works for the sample PDFs I've tried, but
fails on more complex PDFs (as in the parser throws an exception or the
resulting PDF ends up damaged/invalid).

I've not tried parsing/serialising an incrementally updated PDF, but the
current process doesn't attempt to preserve any incremental updates made
to the parsed PDF (if it works at all), instead it would squash/optimize
it into a PDF without incremental updates (single trailer and single
xref table).